Move  functions to librustc_parse
